Move `LocalFeedLoader` and `FeedStore` collaborator to production
